tests: fix multiple library test execution
closes #69
